requirements.txt
streamlit
pandas
plotly
pymysql
SQLAlchemy
cryptography
